scripts/requirements.txt

# Google API client for Gmail attachment crawler
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1
